backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
openai
python-dotenv
alembic
pydantic-settings
psycopg2-binary
